Changed "Payment" package to "payment" Fixed spelling and methods of all the payment.java classes (parent and children) Moved Payment.java into the main package instead of the "payment" package.
