+++ nbproject/private/private.properties
@@ -3,4 +3,4 @@ do.depend=false
 do.jar=true
 javac.debug=true
 javadoc.preview=true
-user.properties.file=/home/pprevide/.netbeans/8.1/build.properties
+user.properties.file=C:\\Users\\user\\AppData\\Roaming\\NetBeans\\8.0.2\\build.properties
