+++ nbproject/private/private.xml
@@ -2,10 +2,6 @@
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/home/pprevide/NetBeansProjects/csc668assignment1/src/csc668assignment1/ProductCatalog.java</file>
-            <file>file:/home/pprevide/NetBeansProjects/csc668assignment1/src/csc668assignment1/Manager.java</file>
-            <file>file:/home/pprevide/NetBeansProjects/csc668assignment1/src/csc668assignment1/Store.java</file>
-        </group>
+        <group/>
     </open-files>
 </project-private>
