+++ src/csc668assignment1/Payment.java
@@ -1,27 +1,32 @@
-package csc668assignment1;
-
 /*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
+ * Only a stub for now
  */
+package csc668assignment1;
 
 /**
  *
- * @author Susanne
+ * @author axelanconaesselmann, Susanne
  */
 public abstract class Payment {
-    private double amountDue;
+    protected double _amountDue;
+    protected String _type;
     
     public Payment(){
-        
+        _amountDue = 0.0;
+        _type = "";
     }
     
-    public Payment(double amountDue){
-        this.amountDue = amountDue;
+    public Payment(double amount) {
+        _amountDue = amount;
     }
     
     public abstract void makePayment(double amountPaid);
-    public abstract double getAmountDue();
+
+    public String getType() {
+        return _type;
+    }
     
+    public double getAmountDue() {
+        return _amountDue;
+    }
 }
