+++ src/csc668assignment1/Transaction.java
@@ -3,8 +3,6 @@
  */
 package csc668assignment1;
 
-import csc668assignment1.Payment.Payment;
-
 /**
  *
  * @author
